pyproject.toml
[project]
name = "gradio-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=5.22.0",
    "isort>=6.0.1",
    "langchain-experimental>=0.3.4",
    "langchain-openai>=0.3.9",
    "langgraph>=0.3.18",
    "matplotlib>=3.10.1",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "pillow>=11.1.0",
    "ruff>=0.11.2",
]
